# sidecar/swaggerui/swagger.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "An API Definition to manage pixiecore hosts",
    "title": "A simple API for PixieCore",
    "contact": {
      "name": "Team Tengu",
      "url": "https://bitbucket.ngage.netapp.com/projects/HCIT/repos/pixiecore-dynamic-rom/browse/sidecar",
      "email": "ng-sf-sre-teamn@netapp.com"
    },
    "version": "0.0.1"
  },
  "basePath": "/",
  "paths": {
    "/about": {
      "get": {
        "description": "Returns information about the application",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "About"
        ],
        "operationId": "About",
        "responses": {
          "200": {
            "description": "About"
          }
        }
      }
    },
    "/history": {
      "get": {
        "description": "Lists booted hosts history",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Inventory"
        ],
        "operationId": "Inventory",
        "responses": {
          "200": {
            "description": "Machine added successfully"
          }
        }
      },
      "post": {
        "description": "Adds a booted host to history",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "Inventory"
        ],
        "operationId": "Inventory",
        "parameters": [
          {
            "description": "Booted machine. Should include a MAC address and architecture",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Machine"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Machine added successfully"
          }
        }
      }
    },
    "/host/{mac}": {
      "get": {
        "description": "Retrieves the boot details of a given host",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "BootConfig"
        ],
        "operationId": "read",
        "parameters": [
          {
            "type": "string",
            "description": "MAC of host config to be returned.",
            "name": "mac",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "MAC configuration not found"
          }
        }
      },
      "post": {
        "description": "Adds a bootable host to the inventory",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "BootConfig"
        ],
        "operationId": "create",
        "parameters": [
          {
            "type": "string",
            "description": "MAC of host config to be added.",
            "name": "mac",
            "in": "path",
            "required": true
          },
          {
            "description": "hostConfig object",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HostConfig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MAC added successfully"
          }
        }
      },
      "delete": {
        "description": "Remove a host from inventory",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "BootConfig"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "type": "string",
            "description": "MAC of host config to be deleted.",
            "name": "mac",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Host removed successfully"
          }
        }
      }
    },
    "/hosts": {
      "get": {
        "description": "Lists all hosts in the inventory",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "BootConfig"
        ],
        "operationId": "Item",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/boot/{mac}": {
      "get": {
        "description": "Retrieves the boot details of a given host",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "BootConfig"
        ],
        "operationId": "boot",
        "parameters": [
          {
            "type": "string",
            "description": "MAC of host config to be returned.",
            "name": "mac",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "MAC details retrieved successfully"
          },
          "404": {
            "description": "MAC not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Host": {
      "description": "Host represents a bootable host",
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/HostConfig"
        },
        "macaddr": {
          "type": "string",
          "x-go-name": "MacAddr"
        }
      },
      "x-go-package": "bitbucket.ngage.netapp.com/scm/hcit/pixiecore-dynamic-rom/sidecar/models"
    },
    "HostConfig": {
      "description": "HostConfig stores boot params for hosts",
      "type": "object",
      "required": [
        "kernel",
        "initrd",
        "cmdline"
      ],
      "properties": {
        "cmdline": {
          "type": "string",
          "x-go-name": "CmdLine",
          "example": "ip=dhcp inst.repo=http://10.117.30.30/images/linux/CentOS-8.1.1911-x86_64-dvd1 inst.ks=https://tengu-boot.s3-us-west-2.amazonaws.com/centos8/centos8.cfg"
        },
        "initrd": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Initrd",
          "example": [
            "http://10.117.30.30/images/linux/CentOS-8.1.1911-x86_64-boot/isolinux/initrd.img"
          ]
        },
        "kernel": {
          "type": "string",
          "x-go-name": "Kernel",
          "example": "http://10.117.30.30/images/linux/CentOS-8.1.1911-x86_64-boot/isolinux/vmlinuz"
        }
      },
      "x-go-package": "bitbucket.ngage.netapp.com/scm/hcit/pixiecore-dynamic-rom/sidecar/models"
    },
    "Machine": {
      "description": "Machine represents a booted host",
      "type": "object",
      "required": [
        "MacAddr"
      ],
      "properties": {
        "Arch": {
          "type": "string",
          "example": "IA32"
        },
        "MacAddr": {
          "type": "string",
          "example": "a1:b2:c3:d4:e5:f6"
        }
      },
      "x-go-package": "bitbucket.ngage.netapp.com/scm/hcit/pixiecore-dynamic-rom/sidecar/models"
    }
  }
}